requirements.txt
# Core ML / Diffusion
# torch>=2.1.0
# diffusers>=0.27.0
# transformers>=4.36.0
# accelerate>=0.25.0
# safetensors>=0.4.0
# peft>=0.7.0

# Image processing
Pillow>=10.0.0

# LLM (Ollama OpenAI-compatible client)
openai>=1.6.0

# Config & Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Web UI (HuggingFace Spaces)
gradio>=4.0.0
